Changed dbsod output type to double

+++ dbsod/dbsod.py
@@ -7,7 +7,7 @@ import numpy as np
 
 dbsod_cpp_path = resources.files('dbsod.build').joinpath('dbsod.so')
 dbsod_cpp = ctypes.CDLL(dbsod_cpp_path)
-dbsod_cpp.dbsod.restype = ctypes.POINTER(ctypes.c_float)
+dbsod_cpp.dbsod.restype = ctypes.POINTER(ctypes.c_double)
 dbsod_cpp.free_array.restype = None
